+++ src/api/routes/user_routes.py
@@ -1,9 +1,9 @@
 from flask import Flask, request, jsonify, url_for, Blueprint
 from api.models import db, User
 
-user_bp = Blueprint('user', __name__)
+user_routes_bp = Blueprint('user_routes_bp', __name__)
 
-@user_bp.route('/hello', methods=['POST', 'GET'])
+@user_routes_bp.route('/hello', methods=['POST', 'GET'])
 def handle_hello():
 
     response_body = {
